Initialize portMappings for existing containers
When RD restarts, we may miss the /tasks/start events for existing containers due to timing issues.
While this might not impact the average user, it can cause noticeable failures during BATs testing.
To address this, we contact the API to get a list of already running containers to initialize the
port mappings, as we will not receive /tasks/start events for these containers again.

Signed-off-by: Nino Kodabande <nkodabande@suse.com>