Create iptables loopback rule for Docker Compose
When a container is created using Docker Compose, the API propagates a map of network names
to network settings. This update checks for the `container.NetworkSettings.Networks` property
to handle network configuration correctly.

Additionally, Docker Compose, by default, creates the following rules in the DOCKER chain:

Also, Docker Compose, by default, creates the following rules in the DOCKER chain:
  DNAT       tcp  --  anywhere             localhost            tcp dpt:80 to:172.18.0.2:80
  DNAT       tcp  --  anywhere             anywhere             tcp dpt:80 to::80

The second rule can be problematic because it uses a wildcard IPv6 address (::), which can
match any incoming TCP traffic destined for port 80. If no service is listening on IPv6,
this can result in a TCP RST (reset) response sent back to the client.

To prevent this issue, we delete the wildcard IPv6 rule before adding our custom rule:
  DNAT       tcp  --  anywhere             anywhere             tcp dpt:80 to:172.18.0.2:80

Note: Even if the `enable_ipv6` property is set to `false` in Docker's Compose configuration,
Docker still creates the wildcard IPv6 rule in iptables. Therefore, we must manually remove
it to avoid this issue.

Signed-off-by: Nino Kodabande <nkodabande@suse.com>