Keep track of loopback iptables rules for deletion
Previously, when we created loopback iptables rules, we abandoned the
rule in the DOCKER chain after container deletion. This happened because
the Docker API did not send port mappings during the container's "die"
event, making it difficult to reconstruct the iptables rule for deletion.

This change caches the iptables deletion commands during container start,
and during the container's "die" event, it looks up the corresponding
deletion command using the container ID as the key in the cached map.

Signed-off-by: Nino Kodabande <nkodabande@suse.com>