WSL: network-setup try to lock OS thread
We're doing some namespace jiggling; sometimes it seems like we can end up
in the wrong namespace (where we fail to configure the veth pair for
veth-rd-ns), and it's in a situation where veth-rd-wsl is visible.
Assuming this is because we've jumped OS threads due to golang green thread
scheduling issues, try to lock the OS thread to mitigate the issue.

Signed-off-by: Mark Yen <mark.yen@suse.com>