Downgrade gvisor-tap-vsock to v0.7.5
As part of the 1.17.0 release, we upgraded the gvisor-tap-vsock library to version v0.8.1.
However, this version has caused issues with DNS resolution behind VPNs and other name resolution problems.
To address these issues, we are planning to release a patch, and this downgrade is part of that solution.

Signed-off-by: Nino Kodabande <nkodabande@suse.com>
(cherry picked from commit 26c3f32321a65e825e9103ee9c5d8996b2daef83)